refactor: split observability config and enablement
